Delete page components
