+++ src/main/java/com/epam/tc/hw4/pages/EpamInformationFrameworkPage.java
@@ -12,8 +12,38 @@ public class EpamInformationFrameworkPage extends BasePage {
         super(webDriver, softAssertions);
     }
 
+    @FindBy(linkText = "HOME")
+    private WebElement elementHomeOnNavigationBar;
+
+    @FindBy(linkText = "CONTACT FORM")
+    private WebElement elementContactsOnNavigationBar;
+
+    @FindBy(linkText = "SERVICE")
+    private WebElement elementServiceOnNavigationBar;
+
+    @FindBy(linkText = "METALS & COLORS")
+    private WebElement elementMetalsAndColorsOnNavigationBar;
+
     @FindBy (css = "li.dropdown:nth-child(1) > a:nth-child(1)")
     private WebElement dropdownButtonWhereLoginFunctions;
+
+    @FindBy(xpath = "//span[text() = 'Home']")
+    private WebElement homeElementOnLeftSideMenu;
+
+    @FindBy(xpath = "//span[text() = 'Service']")
+    private WebElement serviceElementOnLeftSideMenu;
+
+    @FindBy(xpath = "//span[text() = 'Contact form']")
+    private WebElement contactsElementOnLeftSideMenu;
+
+    @FindBy(xpath = "//span[text() = 'Metals & Colors']")
+    private WebElement metalsAndColorsElementOnLeftSideMenu;
+
+    @FindBy(xpath = "//span[text() = 'Elements packs']")
+    private WebElement elementsPacksElementOnLeftSideMenu;
+
+    @FindBy(xpath = "//span[text() = 'Different elements']")
+    private WebElement differentPage;
 
     @FindBy(id = "name")
     private WebElement userNameFieldForLogin;
@@ -95,6 +125,26 @@ public class EpamInformationFrameworkPage extends BasePage {
         softAssertions.assertThat(expectedText).isEqualTo(textUnderBaseIcon.getText());
     }
 
+    @Step("Verify that on header element with name 'HOME' have proper text")
+    public void verifyHomeElementNameOnHeader(String expectedItemText) {
+        softAssertions.assertThat(expectedItemText).isEqualTo(elementHomeOnNavigationBar.getText());
+    }
+
+    @Step ("Verify that on header element with name 'CONTACTS FORM' have proper text")
+    public void verifyContactsElementNameOnHeader(String expectedItemText) {
+        softAssertions.assertThat(expectedItemText).isEqualTo(elementContactsOnNavigationBar.getText());
+    }
+
+    @Step ("Verify that on header element with name 'SERVICE' have proper text")
+    public void verifyServiceElementNameOnHeader(String expectedItemText) {
+        softAssertions.assertThat(expectedItemText).isEqualTo(elementServiceOnNavigationBar.getText());
+    }
+
+    @Step ("Verify that on header element with name 'COLORS & METALS' have proper text")
+    public void verifyColorsAndMetalsElementNameOnHeader(String expectedItemText) {
+        softAssertions.assertThat(expectedItemText).isEqualTo(elementMetalsAndColorsOnNavigationBar.getText());
+    }
+
     @Step ("Verify that under image with name 'Practise' has proper text")
     public void verifyThatProperTextIsUnderPractiseIcon(String expectedText) {
         softAssertions.assertThat(expectedText).isEqualTo(textUnderPractiseIcon.getText());
@@ -119,5 +169,37 @@ public class EpamInformationFrameworkPage extends BasePage {
     public void verifyThatButtonExistOnIframe() {
         webDriver.switchTo().frame("frame");
         softAssertions.assertThat(iframeButton.isDisplayed()).isTrue();
+    }
+
+    @Step("Verify that on left side menu element with name 'Home' have proper text")
+    public void verifyHomeElementNameOnLeftSideMenu(String expectedName) {
+        softAssertions.assertThat(expectedName).isEqualTo(homeElementOnLeftSideMenu.getText());
+    }
+
+    @Step ("Verify that on left side menu element with name 'Service' have proper text")
+    public void verifyServiceElementNameOnLeftSideMenu(String expectedName) {
+        softAssertions.assertThat(expectedName).isEqualTo(serviceElementOnLeftSideMenu.getText());
+    }
+
+    @Step ("Verify that on left side menu element with name 'Contact' have proper text")
+    public void verifyContactElementNameOnLeftSideMenu(String expectedName) {
+        softAssertions.assertThat(expectedName).isEqualTo(contactsElementOnLeftSideMenu.getText());
+    }
+
+    @Step ("Verify that on left side menu element with name 'Metals & Colors' have proper text")
+    public void verifyMetalsAndColorsElementNameOnLeftSideMenu(String expectedName) {
+        softAssertions.assertThat(expectedName).isEqualTo(metalsAndColorsElementOnLeftSideMenu.getText());
+    }
+
+    @Step ("Verify that on left side menu element with name 'Element Packs' have proper text")
+    public void verifyElementPacksElementNameOnLeftSideMenu(String expectedName) {
+        softAssertions.assertThat(expectedName).isEqualTo(elementsPacksElementOnLeftSideMenu.getText());
+    }
+
+    @Step ("Open page with name 'Different Page'")
+    public DifferentPage openDifferentPage() {
+        serviceElementOnLeftSideMenu.click();
+        differentPage.click();
+        return new DifferentPage(webDriver, softAssertions);
     }
 }
