+++ src/test/java/com/epam/tc/hw4/ex2/ExerciseTwoTest.java
@@ -4,7 +4,6 @@ import com.epam.tc.hw4.pages.DifferentPage;
 import com.epam.tc.hw4.pages.EpamInformationFrameworkPage;
 import com.epam.tc.hw4.ConfProperties;
 import com.epam.tc.hw4.TestBase;
-import com.epam.tc.hw4.pages.components.LeftSideMenuComponents;
 import io.qameta.allure.Description;
 import io.qameta.allure.Feature;
 import io.qameta.allure.Story;
@@ -34,7 +33,7 @@ public class ExerciseTwoTest extends TestBase {
         epamInformationFrameworkPage.verifyLoginUser("ROMAN IOVLEV");
 
         //5. Open through the header menu Service -> Different Page
-        DifferentPage differentPage = new LeftSideMenuComponents(webDriver,softAssertions).openDifferentPage();
+        DifferentPage differentPage = epamInformationFrameworkPage.openDifferentPage();
 
         //6.Select checkboxes (Wind,Water)
         differentPage.selectCheckBoxWater();
